fixed many bugs, including the train_stddev not being squared, and just launched the experiment with the exact grad E

+++ refactor_gp/sampling/sample_dae.py
@@ -7,6 +7,9 @@ import refactor_gp
 import refactor_gp.models
 from   refactor_gp.models import dae_untied_weights
 
+import refactor_gp.datasets
+from   refactor_gp.datasets import gaussian_mixture_tools
+
 #import refactor_gp
 import refactor_gp.sampling
 from   refactor_gp.sampling import dispatcher
@@ -27,7 +30,7 @@ def main(argv):
     import json
 
     try:
-        opts, args = getopt.getopt(sys.argv[1:], "hv", ["pickled_dae_dir=", "mcmc_method=", "n_samples=", "thinning_factor=", "burn_in=", "proposal_stddev=", "langevin_stddev=", "langevin_beta=", "output_dir=", "n_E_approx_path=", "want_overview_plots="])
+        opts, args = getopt.getopt(sys.argv[1:], "hv", ["pickled_dae_dir=", "exact_grad_E_from_mixture_mvn_pickle=", "fake_train_stddev=", "mcmc_method=", "n_samples=", "thinning_factor=", "burn_in=", "proposal_stddev=", "langevin_stddev=", "langevin_beta=", "output_dir=", "n_E_approx_path=", "want_overview_plots="])
     except getopt.GetoptError as err:
         # print help information and exit:
         print str(err) # will print something like "option -a not recognized"
@@ -35,6 +38,8 @@ def main(argv):
         sys.exit(2)
 
     pickled_dae_dir = None
+    exact_grad_E_from_mixture_mvn_pickle = None
+    fake_train_stddev = None
     mcmc_method = None
     n_samples = None
     thinning_factor = None
@@ -55,6 +60,10 @@ def main(argv):
             sys.exit()
         elif o in ("--pickled_dae_dir"):
             pickled_dae_dir = a
+        elif o in ("--exact_grad_E_from_mixture_mvn_pickle"):
+            exact_grad_E_from_mixture_mvn_pickle = a
+        elif o in ("--fake_train_stddev"):
+            fake_train_stddev = float(a)
         elif o in ("--mcmc_method"):
             mcmc_method = a
         elif o in ("--n_samples"):
@@ -82,71 +91,115 @@ def main(argv):
     if mcmc_method in ['MH_langevin_grad_E', 'langevin_grad_E']:
         assert not ((langevin_stddev == None) and (langevin_beta == None))
  
-    #print "want_early_termination is %d" % want_early_termination
+    # There are two alternatives to load a DAE.
+    # Either through pickled_dae_dir or exact_grad_E_fake_dae_dir.
+    # ex : 
+    #     pickled_dae_dir = "/data/lisatmp2/alaingui/dae/dae_trained_models/gaussian_mixture_d2/experiment_11/000000"
+    #     exact_grad_E_from_mixture_mvn_pickle = "/data/lisatmp2/alaingui/dae/datasets/gaussian_mixture/d2_eig0.1_comp25_001/train_samples_extra.pkl"
 
-    assert os.path.exists(pickled_dae_dir)
-    assert os.path.isdir(pickled_dae_dir)
+    if pickled_dae_dir is not None:
+        assert os.path.exists(pickled_dae_dir)
+        assert os.path.isdir(pickled_dae_dir)
 
-    pickled_dae_file = os.path.join(pickled_dae_dir, "trained_dae.pkl")
-    pickled_dae_extra_file = os.path.join(pickled_dae_dir, "extra_details.pkl")
-    assert os.path.exists(pickled_dae_file)
-    assert os.path.exists(pickled_dae_extra_file)
+        pickled_dae_file = os.path.join(pickled_dae_dir, "trained_dae.pkl")
+        pickled_dae_extra_file = os.path.join(pickled_dae_dir, "extra_details.pkl")
+        assert os.path.exists(pickled_dae_file)
+        assert os.path.exists(pickled_dae_extra_file)
 
-    # irrelevant values because we load a pickle anyways
-    mydae = dae_untied_weights.DAE_untied_weights(n_inputs = 1,
-                                                  n_hiddens = 1,
-                                                  act_func = ['tanh', 'tanh'])
-    mydae.load_pickle(pickled_dae_file)
-    extra_details = cPickle.load(open(pickled_dae_extra_file, "r"))
+        # irrelevant values because we load a pickle anyways
+        mydae = dae_untied_weights.DAE_untied_weights(n_inputs = 1,
+                                                      n_hiddens = 1,
+                                                      act_func = ['tanh', 'tanh'])
+        mydae.load_pickle(pickled_dae_file)
+        extra_details = cPickle.load(open(pickled_dae_extra_file, "r"))
 
-    # Here's an example of what extra_details contains.
-    """
-    {'act_func': [u'tanh', u'id'],
- 'computational_cost_in_seconds': 373,
- 'early_termination_occurred': True,
- 'lbfgs_rank': 4,
- 'maxiter': 1000,
- 'model_losses': [0.88493400434004688,  0.085861144685915214,  0.014545763882193379,  0.0033357369919251481,  0.0003573931475000403,  3.3169528288705622e-05,  nan,  nan,  nan,  nan],
- 'n_hiddens': 30,
- 'n_inputs': 10,
- 'noise_stddevs': [1.0,  0.2782559402207124,  0.07742636826811268,  0.02154434690031883,  0.005994842503189405,  0.0016681005372000581,  0.00046415888336127757,  0.00012915496650148825,  3.5938136638046215e-05,  9.99999999999998e-06],
- 'train_samples_pickle': '/data/lisatmp2/alaingui/dae/datasets/gaussian_mixture/d10_eig0.1_comp25_001/train_samples.pkl'}
-    """
+        # Here's an example of what extra_details contains.
+        """
+        {'act_func': [u'tanh', u'id'],
+        'computational_cost_in_seconds': 373,
+        'early_termination_occurred': True,
+        'lbfgs_rank': 4,
+        'maxiter': 1000,
+        'model_losses': [0.88493400434004688,  0.085861144685915214,  0.014545763882193379,  0.0033357369919251481,  0.0003573931475000403,  3.3169528288705622e-05,  nan,  nan,  nan,  nan],
+        'n_hiddens': 30,
+        'n_inputs': 10,
+        'noise_stddevs': [1.0,  0.2782559402207124,  0.07742636826811268,  0.02154434690031883,  0.005994842503189405,  0.0016681005372000581,  0.00046415888336127757,  0.00012915496650148825,  3.5938136638046215e-05,  9.99999999999998e-06],
+        'train_samples_pickle': '/data/lisatmp2/alaingui/dae/datasets/gaussian_mixture/d10_eig0.1_comp25_001/train_samples.pkl'}
+        """
 
-    n_inputs = extra_details['n_inputs']
-    n_hiddens = extra_details['n_hiddens']
+        n_inputs = extra_details['n_inputs']
+        n_hiddens = extra_details['n_hiddens']
 
-    # Get the last value for which we don't have 'nan' as value.
-    train_stddev = None
-    for (loss, noise_stddev) in zip(extra_details['model_losses'],
-                                    extra_details['noise_stddevs']):
-        if np.isnan(loss):
-            break
-        else:
-            train_stddev = noise_stddev
-
-    assert train_stddev
-    print "Obtaining grad_E from r and the fact that train_stddev %f" % (train_stddev,)
-
-    dae_params = read_parameters_from_dae(mydae)
-    r = dae_params['r']
-
-    # don't forget the minus sign here !
-    # remember that r(x) - x is propto  -1 * dE/dx
-    grad_E = lambda x: -1*(r(x)-x) / train_stddev
-
-    # We always sample from the origin.
-    x0 = np.zeros((n_inputs,))
+        # Get the last value for which we don't have 'nan' as value.
+        train_stddev = None
+        for (loss, noise_stddev) in zip(extra_details['model_losses'],
+                                        extra_details['noise_stddevs']):
+            if np.isnan(loss):
+                break
+            else:
+                train_stddev = noise_stddev
 
 
-    sampling_options = {'x0':x0,
-                        'f_prime':dae_params['f_prime'],
-                        'r':dae_params['r'],
-                        'r_prime':dae_params['r_prime'],
-                        'mcmc_method':mcmc_method,
-                        'grad_E':grad_E,
-                        'n_samples':n_samples,
-                        'train_stddev':train_stddev}
+        assert train_stddev
+        print "Obtaining grad_E from r and the fact that train_stddev %f" % (train_stddev,)
+
+        dae_params = read_parameters_from_dae(mydae)
+        r = dae_params['r']
+
+        # don't forget the minus sign here !
+        # remember that r(x) - x is propto  -1 * dE/dx
+        def grad_E(x):
+            return -1*(r(x)-x) / train_stddev**2
+
+        x0 = np.zeros((n_inputs,))
+        sampling_options = {'x0':x0,
+                            'f_prime':dae_params['f_prime'],
+                            'r':dae_params['r'],
+                            'r_prime':dae_params['r_prime'],
+                            'grad_E':grad_E,
+                            'train_stddev':train_stddev}
+
+    elif exact_grad_E_from_mixture_mvn_pickle is not None:
+        assert os.path.exists(exact_grad_E_from_mixture_mvn_pickle)
+        assert fake_train_stddev is not None
+        assert fake_train_stddev > 0
+
+        extra_props = cPickle.load(open(exact_grad_E_from_mixture_mvn_pickle, "r"))
+        # This nets us a structure that looks like the following.
+        #   extra_props = {'component_means':component_means,
+        #                  'component_covariances':component_covariances,
+        #                  'd':d,
+        #                  'ratio_eigvals':ratio_eigvals,
+        #                  'n_components':n_components,
+        #                  'f_parameters':f_parameters,
+        #                  'computational_cost_in_seconds':computational_cost_in_seconds}
+
+        def grad_E(x):
+            return gaussian_mixture_tools.grad_E(x, extra_props['component_means'], extra_props['component_covariances'])
+
+        def r(x):
+            return x + fake_train_stddev**2 * grad_E(x)
+
+        def r_prime(x):
+            return np.eye(x.shape[0])
+
+        # We always sample from the origin and this is fine
+        # because we generated the data so that would have 
+        # some density at the origins.
+        x0 = np.zeros((extra_props['d'],))
+
+        sampling_options = {'x0':x0,
+                            'r':r,
+                            'r_prime':r_prime,
+                            'grad_E':grad_E,
+                            'train_stddev':fake_train_stddev}
+
+    else:
+        raise("You have to use either the --pickled_dae_dir or --exact_grad_E_from_mixture_mvn_pickle option.")
+
+
+    sampling_options['mcmc_method'] = mcmc_method
+    sampling_options['n_samples'] = n_samples
 
     if burn_in is not None:
         sampling_options['burn_in'] = burn_in
@@ -192,39 +245,34 @@ def main(argv):
     # Get rid of the memory-intensive 'samples'
     # field because it was already written in mcmc_samples.pkl.
     del sampling_extra_details['samples']
-    sampling_extra_details['pickled_dae_dir'] = pickled_dae_dir
-    sampling_extra_details['mcmc_method'] = mcmc_method
-    sampling_extra_details['n_samples'] = n_samples
-    sampling_extra_details['thinning_factor'] = thinning_factor
-    sampling_extra_details['burn_in'] = burn_in
-    sampling_extra_details['proposal_stddev'] = proposal_stddev
-    sampling_extra_details['langevin_stddev'] = langevin_stddev
-    sampling_extra_details['langevin_beta'] = langevin_beta
-    sampling_extra_details['train_stddev'] = train_stddev
-    sampling_extra_details['n_E_approx_path'] = n_E_approx_path
+    for (k,v) in sampling_options.items():
+        assert not sampling_extra_details.has_key(k)
+        if k not in ['r', 'r_prime', 'f', 'f_prime', 'grad_E', 'x0']:
+            sampling_extra_details[k] = v
+        
 
     sampling_extra_pickle_file = os.path.join(output_dir, "sampling_extra_details.pkl")
     sampling_extra_json_file = os.path.join(output_dir, "sampling_extra_details.json")
     cPickle.dump(sampling_extra_details, open(sampling_extra_pickle_file, "w"))
-    json.dump(sampling_extra_details, open(sampling_extra_json_file, "w"))
+    json.dump(sampling_extra_details, open(sampling_extra_json_file, "w"), sort_keys=True, indent=4, separators=(',', ': '))
     print "Wrote %s" % (sampling_extra_pickle_file,)
     print "Wrote %s" % (sampling_extra_json_file,)
 
 
-    #### Copy the original DAE. ####
+    #### Copy the original DAE (if applicable). ####
 
     # We'll copy the original DAE here because we
     # want to make sure that it stays around when
     # we want to inspect the samples and check the
     # characteristics of the DAE.
 
-    import shutil
-    output_dir_subdir = os.path.join(output_dir, "trained_dae")
-    if os.path.exists(output_dir_subdir):
-        shutil.rmtree(output_dir_subdir)
-    shutil.copytree(pickled_dae_dir, output_dir_subdir)
-    print "Transfered trained DAE from %s to %s" % (pickled_dae_dir, output_dir_subdir)
-
+    if pickled_dae_dir is not None:
+        import shutil
+        output_dir_subdir = os.path.join(output_dir, "trained_dae")
+        if os.path.exists(output_dir_subdir):
+            shutil.rmtree(output_dir_subdir)
+            shutil.copytree(pickled_dae_dir, output_dir_subdir)
+            print "Transfered trained DAE from %s to %s" % (pickled_dae_dir, output_dir_subdir)
 
     #### Option to generate plots. ####
